+++ src/btech/movement/ds_bay_api.h
@@ -8,6 +8,6 @@ typedef struct BattleMap BattleMap;
 void mech_createbays(DbRef player, void *data, char *buffer);
 int dropship_bay_number(Mech *ds, int direction);
 int dropship_bay_direction(Mech *ds, int bay);
-int dropship_bay_in_adjacent_hex(Mech *seer, Mech *dropship, long *bay);
+int dropship_bay_in_adjacent_hex(Mech *seer, Mech *dropship, int *bay);
 void mech_enterbay(DbRef player, void *data, char *buffer);
 int dropship_leave(BattleMap *map, Mech *mech);
